+++ R/var.relations.R
@@ -43,14 +43,15 @@
 #'
 #' @export
 var.relations <- function(x = NULL, y = NULL, num.trees = 500, type = "regression", s = NULL, mtry = NULL, min.node.size = 1,
-                          num.threads = NULL, status = NULL, save.ranger = FALSE, create.forest = TRUE, forest = NULL,
+                          num.threads = NULL, status = NULL, save.ranger = FALSE, create.forest = is.null(forest), forest = NULL,
                           save.memory = FALSE, case.weights = NULL,
                           variables, candidates, t = 5, select.rel = TRUE) {
-  if (!is.data.frame(x)) {
-    stop("x has to be a data frame")
-  }
   if (create.forest) {
     ## check data
+    if (!is.data.frame(x)) {
+      stop("x has to be a data frame")
+    }
+
     if (length(y) != nrow(x)) {
       stop("length of y and number of rows in x are different")
     }
@@ -123,25 +124,24 @@ var.relations <- function(x = NULL, y = NULL, num.trees = 500, type = "regressio
     forest <- list(trees = trees.surr, allvariables = colnames(data[, -1]))
   }
 
-  if (!create.forest) {
-    if (is.null(forest)) {
-      stop("set create.forest to TRUE or analyze an existing random forest specified by parameter forest")
-    }
+  if (!create.forest && is.null(forest)) {
+    stop("set create.forest to TRUE or analyze an existing random forest specified by parameter forest")
   }
+
   trees <- forest[["trees"]]
   allvariables <- forest[["allvariables"]]
 
-  if (all(candidates %in% allvariables)) {
-    if (all(variables %in% allvariables)) {
-      # count surrogates
-      s <- count.surrogates(trees)
-      results.meanAdjAgree <- meanAdjAgree(trees, variables, allvariables, candidates, t = t, s$s.a, select.var = select.rel, num.threads = num.threads)
-    } else {
-      stop("allvariables do not contain the chosen variables")
-    }
-  } else {
+  if (!all(candidates %in% allvariables)) {
     stop("allvariables do not contain the candidate variables")
   }
+  if (!all(variables %in% allvariables)) {
+    stop("allvariables do not contain the chosen variables")
+  }
+
+  # count surrogates
+  s <- count.surrogates(trees)
+  results.meanAdjAgree <- meanAdjAgree(trees, variables, allvariables, candidates, t = t, s$s.a, select.var = select.rel, num.threads = num.threads)
+
   if (select.rel) {
     surr.var <- results.meanAdjAgree$surr.var
     varlist <- list()